+++ analytics-provider/src/main/java/org/digma/intellij/plugin/analytics/RestAnalyticsProvider.java
@@ -336,6 +336,16 @@ public class RestAnalyticsProvider implements AnalyticsProvider, Closeable {
     }
 
     @Override
+    public String getHighlightsScaling(HighlightsRequest request) {
+        return execute(() -> client.analyticsProvider.getHighlightsScaling(request));
+    }
+
+    @Override
+    public String getSpanInfo(String spanCodeObjectId) {
+        return execute(() -> client.analyticsProvider.getSpanInfo(spanCodeObjectId));
+    }
+
+    @Override
     public HttpResponse lowLevelCall(HttpRequest request) {
 
         okhttp3.Response okHttp3Response;
@@ -954,5 +964,18 @@ public class RestAnalyticsProvider implements AnalyticsProvider, Closeable {
         @POST("/authentication/register")
         Call<String> register(@Body Map<String, Object> request);
 
+        @Headers({
+                "Accept: application/+json",
+                "Content-Type:application/json"
+        })
+        @POST("/highlights/scaling")
+        Call<String> getHighlightsScaling(@Body HighlightsRequest request);
+
+        @Headers({
+                "Accept: application/+json",
+                "Content-Type:application/json"
+        })
+        @GET("/spans/info")
+        Call<String> getSpanInfo(@Query("SpanCodeObjectId") String spanCodeObjectId);
     }
 }