+++ jvm-common/src/main/kotlin/org/digma/intellij/plugin/idea/execution/OTELJarProvider.kt
@@ -8,16 +8,15 @@ import java.io.FileOutputStream
 import java.net.URL
 import java.nio.file.Files
 import java.nio.file.StandardCopyOption
+import java.util.Properties
 import java.util.concurrent.locks.ReentrantLock
 import kotlin.io.path.deleteIfExists
 
 private const val JARS_DIR_PREFIX = "digma-otel-jars"
 private const val RESOURCE_LOCATION = "otelJars"
-private const val OTEL_AGENT_JAR_URL =
-//  "https://github.com/open-telemetry/opentelemetry-java-instrumentation/releases/latest/download/opentelemetry-javaagent.jar"
-    "https://github.com/open-telemetry/opentelemetry-java-instrumentation/releases/download/v2.1.0/opentelemetry-javaagent.jar"
-private const val DIGMA_AGENT_EXTENSION_JAR_URL =
-    "https://github.com/digma-ai/otel-java-instrumentation/releases/latest/download/digma-otel-agent-extension.jar"
+
+private const val OTEL_AGENT_JAR_PROP_NAME = "otel-agent"
+private const val DIGMA_EXTENSION_JAR_PROP_NAME = "digma-extension"
 private const val OTEL_AGENT_JAR_NAME = "opentelemetry-javaagent.jar"
 private const val DIGMA_AGENT_EXTENSION_JAR_NAME = "digma-otel-agent-extension.jar"
 
@@ -32,10 +31,15 @@ class OTELJarProvider {
 
     private val downloadDir: File = File(System.getProperty("java.io.tmpdir"), JARS_DIR_PREFIX)
 
+    private val jarsUrls = Properties()
+
     private val lock = ReentrantLock()
 
 
     init {
+
+        jarsUrls.load(this::class.java.getResourceAsStream("/jars-urls.properties"))
+
         //unpack and download on service initialization.
         //will happen per IDE session
         Thread {
@@ -137,8 +141,8 @@ class OTELJarProvider {
         val runnable = Runnable {
 
             try {
-                val otelUrl = System.getProperty("org.digma.otel.agentUrl", OTEL_AGENT_JAR_URL)
-                val extensionUrl = System.getProperty("org.digma.otel.extensionUrl", DIGMA_AGENT_EXTENSION_JAR_URL)
+                val otelUrl = System.getProperty("org.digma.otel.agentUrl", jarsUrls.getProperty(OTEL_AGENT_JAR_PROP_NAME))
+                val extensionUrl = System.getProperty("org.digma.otel.extensionUrl", jarsUrls.getProperty(DIGMA_EXTENSION_JAR_PROP_NAME))
 
                 downloadAndCopyJar(URL(otelUrl), getOtelAgentJar())
                 downloadAndCopyJar(URL(extensionUrl), getDigmaAgentExtensionJar())