+++ jvm-common/src/main/kotlin/org/digma/intellij/plugin/idea/execution/flavor/BaseInstrumentationFlavor.kt
@@ -52,6 +52,7 @@ abstract class BaseInstrumentationFlavor : InstrumentationFlavor {
 
     open fun isTest(
         instrumentationService: RunConfigurationInstrumentationService,
+        parametersExtractor: ParametersExtractor,
         configuration: RunConfiguration,
         params: SimpleProgramParameters
     ): Boolean {