+++ jvm-common/src/main/kotlin/org/digma/intellij/plugin/idea/execution/flavor/JavaServerInstrumentationFlavor.kt
@@ -35,6 +35,7 @@ class JavaServerInstrumentationFlavor : DefaultInstrumentationFlavor() {
 
     override fun isTest(
         instrumentationService: RunConfigurationInstrumentationService,
+        parametersExtractor: ParametersExtractor,
         configuration: RunConfiguration,
         params: SimpleProgramParameters
     ): Boolean {