+++ src/main/kotlin/org/digma/intellij/plugin/ui/highlights/HighlightsService.kt
@@ -83,4 +83,28 @@ class HighlightsService(val project: Project) : InsightsServiceImpl(project) {
             null
         }
     }
+
+    fun getHighlightsScaling(request: HighlightsRequest): String? {
+        EDT.assertNonDispatchThread()
+
+        return try {
+            val highlightsPerformance = AnalyticsService.getInstance(project).getHighlightsScaling(request)
+            highlightsPerformance
+        } catch (e: AnalyticsServiceException) {
+            Log.warnWithException(logger, project, e, "Error loading highlights top insights {}", e.message)
+            null
+        }
+    }
+
+    fun getSpanInfo(spanCodeObjectId: String): String? {
+        EDT.assertNonDispatchThread()
+
+        return try {
+            val spanInfo = AnalyticsService.getInstance(project).getSpanInfo(spanCodeObjectId)
+            spanInfo
+        } catch (e: AnalyticsServiceException) {
+            Log.warnWithException(logger, project, e, "Error loading highlights top insights {}", e.message)
+            null
+        }
+    }
 }