+++ CHANGELOG.md
@@ -1,13 +1,40 @@
 # Changelog
 
-## [2.0.330] - 2024-05-06
+## [2.0.334] - 2024-05-15
+
+- Updated ui by @opoliarush in https://github.com/digma-ai/digma-intellij-plugin/pull/2158
+- Updated ui by @opoliarush in https://github.com/digma-ai/digma-intellij-plugin/pull/2159
+- catch error in org.digma.intellij.plugin.analytics.BackendInfoHolder.… by @shalom938 in https://github.com/digma-ai/digma-intellij-plugin/pull/2160
+
+## 2.0.333 - 2024-05-13
+
+- Update UI by @opoliarush in https://github.com/digma-ai/digma-intellij-plugin/pull/2151
+- change AuthenticationException log level to debug by @shalom938 in https://github.com/digma-ai/digma-intellij-plugin/pull/2154
+- force observability with DIGMA_OBSERVABILITY #Closes 2144 by @shalom938 in https://github.com/digma-ai/digma-intellij-plugin/pull/2156
+- Added span info endpoint by @opoliarush in https://github.com/digma-ai/digma-intellij-plugin/pull/2157
+
+## 2.0.332 - 2024-05-10
+
+- Added scaling highlights by @opoliarush in https://github.com/digma-ai/digma-intellij-plugin/pull/2135
+- Update UI by @opoliarush in https://github.com/digma-ai/digma-intellij-plugin/pull/2149
+- fix-CycleInitializationException by @shalom938 in https://github.com/digma-ai/digma-intellij-plugin/pull/2150
+
+## 2.0.331 - 2024-05-09
+
+- clear current env when deleting last env Closes #2139 by @shalom938
+  in https://github.com/digma-ai/digma-intellij-plugin/pull/2140
+- put jars urls in properties file by @shalom938 in https://github.com/digma-ai/digma-intellij-plugin/pull/2147
+
+## 2.0.330 - 2024-05-06
 
 - Update Digmathon by @kshmidt-digma in https://github.com/digma-ai/digma-intellij-plugin/pull/2134
 
 ## 2.0.329 - 2024-05-05
 
-- prevent posthog errors when can't connect Closes #2041 by @shalom938 in https://github.com/digma-ai/digma-intellij-plugin/pull/2137
-- don't log NoSelectedEnvironmentException in warn level Closes #2136 by @shalom938 in https://github.com/digma-ai/digma-intellij-plugin/pull/2138
+- prevent posthog errors when can't connect Closes #2041 by @shalom938
+  in https://github.com/digma-ai/digma-intellij-plugin/pull/2137
+- don't log NoSelectedEnvironmentException in warn level Closes #2136 by @shalom938
+  in https://github.com/digma-ai/digma-intellij-plugin/pull/2138
 
 ## 2.0.328 - 2024-05-01